start splitting things into different modules/packages
instead of one big ol long file, we'll have several files. and I'll set
them up for dependency injection too
