+++ caseta_to_mqtt/caseta/__init__.py
@@ -1,0 +1,6 @@
+from datetime import timedelta
+
+
+DOUBLE_CLICK_WINDOW = timedelta(milliseconds=500)
+BUTTON_WATCHER_SLEEP_DURATION = timedelta(milliseconds=250)
+BUTTON_WATCHER_MAX_DURATION = timedelta(seconds=5)
