+++ caseta_to_mqtt/z2m/publisher.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+import asyncio
+import json
+import logging
+
+import aiomqtt
+
+from caseta_to_mqtt.z2m.model import Zigbee2MqttGroup
+
+
+LOGGER = logging.getLogger(__name__)
+
+
+class Zigbee2mqttPublisher:
+    def __init__(self, mqtt_client: aiomqtt.Client) -> Zigbee2mqttPublisher:
+        self._mqtt_client = mqtt_client
+
+    async def turn_on_group(self, group: Zigbee2MqttGroup):
+        async with self._mqtt_client as client:
+            await client.publish(group.topic, json.dumps({"on": True}))
+
+    async def turn_off_group(self, group: Zigbee2MqttGroup):
+        async with self._mqtt_client as client:
+            await client.publish(group.topic, json.dumps({"on": False}))
+
+    async def publish_loop(self):
+        while True:
+            LOGGER.info("sleeping, then turning on")
+            await asyncio.sleep(2)
+            # await self.turn_on_group(group)
+            LOGGER.info("sleeping, then turning off")
+            await asyncio.sleep(2)
+            # await self.turn_off_group(group)
